Bugfix: response bodies not closed when retrying requests
Usually, when requests fail and are to be retried, there is no need to
close the response body because there is no body. However, if there is a
response body, it was not closed, which could lead to connection leaks.

This commit fixes the issue by closing the body between attempts, if
there is a response.